Removed unnesessary overhead during clip recovery
During clip recovery creating double the links is unnecessary because the second half exceeds the duration of the stream. It is still accounted that only even numbered links are beeing added to the list. The total number of links to check is half of the duration in seconds. Therefore the counter displayed still tracks seconds not links.